docs: remove GH pages repo as submodule, adjust publish.sh
Would need to bump zrepl main repo for every publish otherwise...

+++ docs/publish.sh
@@ -2,24 +2,52 @@
 set -eo pipefail
 
 GHPAGESREPO="git@github.com:zrepl/zrepl.github.io.git"
+SCRIPTDIR=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )
+PUBLICDIR="${SCRIPTDIR}/public"
+
+checkout_repo_msg() {
+    echo "checkout ${GHPAGESREPO} to ${PUBLICDIR}:"
+    echo "	git clone ${GHPAGESREPO} ${PUBLICDIR}"
+}
 
 exit_msg() {
     echo "error, exiting..."
 }
 trap exit_msg EXIT
 
+cd "$SCRIPTDIR"
+
+if [ ! -d "$PUBLICDIR" ]; then
+    checkout_repo_msg
+    exit 1
+fi
+
 echo -n "PRESS ENTER to confirm you commited the docs changes to the zrepl repo"
 read
 
-cd public
+pushd "$PUBLICDIR" 
+
 echo "verify we're in the GitHub pages repo..."
 git remote get-url origin | grep -E "^${GHPAGESREPO}\$"
+if [ "$?" -ne "0" ] ;then 
+    checkout_repo_msg
+    exit 1
+fi
+
+echo "resetting GitHub pages repo to latest commit"
+git fetch origin
+git reset --hard origin/master
+
 echo "cleaning GitHub pages repo"
 git clean -dn
-cd ..
+
+popd
+
 echo "building site"
 hugo
-cd public
+
+pushd "$PUBLICDIR"
+
 echo "adding and commiting all changes in GitHub pages repo"
 git add -A
 git commit -m "hugo render from publish.sh - `date -u`"
